DHD: 4390 - For MSI case move CFG INTSTATUS (for CTO) reading to DPC instead of ISR
This change is based on the below customer request:
Based on recent testing on P24, we found that system performance is
highly impacted by the PCIE read, per shared in #comment9.
we really need to evaluate the feasibility of moving the
interrupt status read work from the hardirq handler into a
softirq handler. Otherwise, time for PCIE to resume to L0 is
introducing delays to other tasks. especially causing delays in system
performance testing when wifi was running light traffics.

For MSI, we are already skipping reading chip INTSTATUS.
The cfg INTSTATUS also will be moved to DPC for MSI case.
For INTx case the status-quo is maintained.

Bug: 283029172

Change-Id: I67248a3f8b3b7549c1eac9b56f617358a629a031
Signed-off-by: Yadukishore Kodicherla <yadukishore.kodicherla@broadcom.corp-partner.google.com>