bcmdhd: Fixed to set the pwrcap to MAX value when the framework give DHD the CELLAVOID_NO_POWER_CAP.
Issue:
On the framework side, when they set some channels to unsafe in order to avoid these channels due to cellular coex, the upper layer only sets frequencies/bands to avoid.
According to GG, in the current API, there is no param for pwrcap.
Although the framework doesn't deliver any pwrcap value, DHD sets power of these channels to -1.

Root Cause:
Currently, the framework gives DHD the pwrcap to 2147483647(0x7FFF_FFFF) as default value.
In the HAL API format, pwrcap is defined as signed 32 bit. On the other hand pwrcap is defined as signed 8 bit in DHD API format.
Due to wrong typecasting, pwrcap was set to -1(0xFF).

Fix:
Fixed to set the pwrcap to MAX value when the framework give DHD the pwrcap to CELLAVOID_NO_POWER_CAP(0x7FFF_FFFF).
Also, in the framework side, pwrcap is the unit of dBm not, qdBm. Therefore we fixed MIN/MAX value to dBm scale.

Bug: 341900124
Test: Verified with Pixel device
Test: Regression test

Change-Id: Ie7ad9f0bdebfa8973f43ea3190a99848dbf67ed8
Signed-off-by: Dewey Lee <dewey.lee@broadcom.corp-partner.google.com>